+++ components/outpost/LoadingOutposts.tsx
@@ -1,4 +1,4 @@
-import { OutpostCardSkeleton } from "./OutpostCardSkeleton";
+import { OutpostCardSkeleton } from "./outpostCard/skeleton";
 
 interface LoadingOutpostsProps {
   count?: number;
