+++ components/outpost/outpostCard/skeleton/index.tsx
